+++ pom.xml
@@ -4,10 +4,10 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.aman.metrics</groupId>
-    <artifactId>async-codahale</artifactId>
+    <groupId>com.github.isopropylcyanide</groupId>
+    <artifactId>async-codahale-metrics</artifactId>
     <version>1.0-SNAPSHOT</version>
-    <name>Async Metrics</name>
+    <name>Async Codahale Metrics</name>
     <description>Leverage metrics for asynchronously returning methods</description>
     <url>https://github.com/isopropylcyanide/async-metrics-codahale</url>
