+++ src/main/java/com/github/isopropylcyanide/asyncmetrics/module/AsyncMetricsModule.java
@@ -11,15 +11,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.async.metrics.module;
+package com.github.isopropylcyanide.asyncmetrics.module;
 
-import com.async.metrics.annotation.AsyncExceptionMetered;
-import com.async.metrics.annotation.AsyncMetered;
-import com.async.metrics.annotation.AsyncTimed;
-import com.async.metrics.interceptor.AsyncExceptionMeteredInterceptor;
-import com.async.metrics.interceptor.AsyncMeteredInterceptor;
-import com.async.metrics.interceptor.AsyncTimedInterceptor;
 import com.codahale.metrics.MetricRegistry;
+import com.github.isopropylcyanide.asyncmetrics.annotation.AsyncExceptionMetered;
+import com.github.isopropylcyanide.asyncmetrics.annotation.AsyncMetered;
+import com.github.isopropylcyanide.asyncmetrics.annotation.AsyncTimed;
+import com.github.isopropylcyanide.asyncmetrics.interceptor.AsyncExceptionMeteredInterceptor;
+import com.github.isopropylcyanide.asyncmetrics.interceptor.AsyncMeteredInterceptor;
+import com.github.isopropylcyanide.asyncmetrics.interceptor.AsyncTimedInterceptor;
 import com.google.inject.AbstractModule;
 import com.google.inject.matcher.Matchers;
