+++ src/test/java/com/github/isopropylcyanide/asyncmetrics/interceptor/AsyncMeteredAspectTest.java
@@ -11,11 +11,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.async.metrics.interceptor;
+package com.github.isopropylcyanide.asyncmetrics.interceptor;
 
-import com.async.metrics.annotation.AsyncTimed;
+import com.codahale.metrics.Meter;
 import com.codahale.metrics.MetricRegistry;
-import com.codahale.metrics.Timer;
+import com.github.isopropylcyanide.asyncmetrics.annotation.AsyncMetered;
 import org.aopalliance.intercept.MethodInvocation;
 import org.junit.Before;
 import org.junit.Test;
@@ -28,13 +28,14 @@ import java.util.concurrent.CompletableFuture;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
+import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @SuppressWarnings({"unchecked", "WeakerAccess"})
-public class AsyncTimedAspectTest {
+public class AsyncMeteredAspectTest {
 
     @Mock
     private MetricRegistry metricRegistry;
@@ -42,37 +43,36 @@ public class AsyncTimedAspectTest {
     @Mock
     private MethodInvocation invocation;
 
-    private AsyncTimedInterceptor interceptor;
+    private AsyncMeteredInterceptor interceptor;
 
     @Before
     public void setUp() {
         MockitoAnnotations.initMocks(this);
-        this.interceptor = new AsyncTimedInterceptor(metricRegistry);
+        this.interceptor = new AsyncMeteredInterceptor(metricRegistry);
     }
 
-    @AsyncTimed(name = "N1")
+    @AsyncMetered(name = "N1")
     public void testMethodWithAnnotationWithName() {
     }
 
-    @AsyncTimed
+    @AsyncMetered
     public void testMethodWithAnnotationWithNoName() {
     }
 
     @Test
     public void testAdviceMarksMetricsWhenSystemPropertyIsNotPresentAndAnnotationHasName() throws Throwable {
         Method method = this.getClass().getMethod("testMethodWithAnnotationWithName");
-        Timer timer = mock(Timer.class);
-        Timer.Context context = mock(Timer.Context.class);
+        Meter meter = mock(Meter.class);
 
         when(invocation.getMethod()).thenReturn(method);
-        when(metricRegistry.timer(Mockito.contains("N1"))).thenReturn(timer);
-        when(timer.time()).thenReturn(context);
+        when(metricRegistry.meter(Mockito.contains("N1"))).thenReturn(meter);
+        doNothing().when(meter).mark();
 
         when(invocation.proceed()).thenReturn(CompletableFuture.completedFuture("Done"));
         CompletableFuture<String> proceed = (CompletableFuture) interceptor.invoke(invocation);
 
         verify(invocation, times(1)).proceed();
-        verify(context, times(1)).stop();
+        verify(meter, times(1)).mark();
         assertTrue(proceed.isDone());
         assertEquals("Done", proceed.get());
     }
@@ -80,18 +80,17 @@ public class AsyncTimedAspectTest {
     @Test
     public void testAdviceMarksMetricsWhenSystemPropertyIsNotPresentAndAnnotationHasNoName() throws Throwable {
         Method method = this.getClass().getMethod("testMethodWithAnnotationWithNoName");
-        Timer timer = mock(Timer.class);
-        Timer.Context context = mock(Timer.Context.class);
+        Meter meter = mock(Meter.class);
 
         when(invocation.getMethod()).thenReturn(method);
-        when(metricRegistry.timer(Mockito.contains("testMethodWithAnnotationWithNoName"))).thenReturn(timer);
-        when(timer.time()).thenReturn(context);
+        when(metricRegistry.meter(Mockito.contains("testMethodWithAnnotationWithNoName"))).thenReturn(meter);
+        doNothing().when(meter).mark();
 
         when(invocation.proceed()).thenReturn(CompletableFuture.completedFuture("Done"));
         CompletableFuture<String> proceed = (CompletableFuture) interceptor.invoke(invocation);
 
         verify(invocation, times(1)).proceed();
-        verify(context, times(1)).stop();
+        verify(meter, times(1)).mark();
         assertTrue(proceed.isDone());
         assertEquals("Done", proceed.get());
     }
